Add files via upload
I've made this in the format of original NGPR implementation code (ie, it fits there in the place of likelihood function), but the process itself can be taken to fit it any of your problems. Its based on first computing eigen values of Kernel Matrix used in NLML and then using that as a way to approximate noise.